# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: E-Wallet Digital Payment Service API
  description: |
    API Documentation for Digital Payment Service (E-Wallet) with Microservices Architecture.
    
    This API provides endpoints for user management, wallet operations, and transaction processing.
  version: 1.0.0
  contact:
    name: E-Wallet API Support

servers:
  - url: http://localhost:3000/api
    description: API Gateway (Development)

tags:
  - name: Users
    description: User authentication and management
  - name: Wallets
    description: Wallet operations and balance management
  - name: Transactions
    description: Payment and transaction history

paths:
  # User Service Endpoints
  /users/register:
    post:
      tags:
        - Users
      summary: Register a new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: alice
                email:
                  type: string
                  format: email
                  example: alice@example.com
                password:
                  type: string
                  format: password
                  example: password123
                fullName:
                  type: string
                  example: Alice Johnson
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      username:
                        type: string
                        example: alice
                      email:
                        type: string
                        example: alice@example.com
                      fullName:
                        type: string
                        example: Alice Johnson
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '409':
          description: Username or email already exists
        '400':
          description: Invalid input

  /users/login:
    post:
      tags:
        - Users
      summary: Login user
      description: Authenticate user and get access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: alice
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      username:
                        type: string
                        example: alice
                      email:
                        type: string
                        example: alice@example.com
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials

  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Get authenticated user's profile information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  username:
                    type: string
                    example: alice
                  email:
                    type: string
                    example: alice@example.com
                  full_name:
                    type: string
                    example: Alice Johnson
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized

  # Wallet Service Endpoints
  /wallets/balance:
    get:
      tags:
        - Wallets
      summary: Get wallet balance
      description: Get current balance for authenticated user
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    example: 1
                  balance:
                    type: number
                    format: decimal
                    example: 100000.00
                  currency:
                    type: string
                    example: IDR
        '401':
          description: Unauthorized

  /wallets/topup:
    post:
      tags:
        - Wallets
      summary: Top up wallet
      description: Add funds to user's wallet
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  format: decimal
                  minimum: 0.01
                  example: 50000
      responses:
        '200':
          description: Top up successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Top up successful
                  user_id:
                    type: integer
                    example: 1
                  new_balance:
                    type: number
                    format: decimal
                    example: 150000.00
        '400':
          description: Invalid amount

  # Transaction Service Endpoints
  /transactions/payment:
    post:
      tags:
        - Transactions
      summary: Send payment
      description: Create a payment transaction to another user
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
          description: User ID (sender)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipientId
                - amount
              properties:
                recipientId:
                  type: integer
                  example: 2
                amount:
                  type: number
                  format: decimal
                  minimum: 0.01
                  example: 50000
                description:
                  type: string
                  example: Payment for services
      responses:
        '201':
          description: Payment successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Payment successful
                  transaction:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      sender_id:
                        type: integer
                        example: 1
                      recipient_id:
                        type: integer
                        example: 2
                      amount:
                        type: number
                        format: decimal
                        example: 50000.00
                      description:
                        type: string
                        example: Payment for services
                      status:
                        type: string
                        example: completed
                      created_at:
                        type: string
                        format: date-time
        '400':
          description: Invalid request or insufficient balance
        '404':
          description: Recipient not found

  /transactions/history:
    get:
      tags:
        - Transactions
      summary: Get transaction history
      description: Get transaction history for authenticated user
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
          description: User ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of transactions to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of transactions to skip
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    example: 1
                  transactions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        sender_id:
                          type: integer
                        recipient_id:
                          type: integer
                        amount:
                          type: number
                          format: decimal
                        description:
                          type: string
                        status:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                    example: 10

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT